refactor: replace debug() with Logging.utils; split implementation into separate files
