+++ src/hw1/ParallelExternalLongSorter.java
@@ -5,48 +5,55 @@
 //region imports
 package hw1;
 
+import java.io.IOException;
 import java.nio.Buffer;
 import java.nio.LongBuffer;
 import java.nio.channels.FileChannel;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.*;
-import java.util.concurrent.Callable;
+import java.util.List;
+import java.util.PriorityQueue;
+import java.util.Set;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
+import java.util.logging.Logger;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
+import static hw1.Utils.*;
 import static java.nio.channels.FileChannel.MapMode.READ_ONLY;
 import static java.nio.channels.FileChannel.MapMode.READ_WRITE;
 import static java.nio.file.StandardOpenOption.*;
 //endregion
 
-public class ParallelExternalLongSorter {
-
+@SuppressWarnings("ClassCanBeRecord") // suppressed because colab notebook's version of java doesn't support 'record'
+class ParallelExternalLongSorter {
     //region constants
     static final int THREADPOOL_TIMEOUT_SECONDS = 60;
-    static final String DEFAULT_INPUT_FILENAME = "array.bin";
-    static final String DEFAULT_OUTPUT_FILENAME = "sorted.bin";
-    static final int DEFAULT_NTHREADS = Runtime.getRuntime().availableProcessors();
     private static final int BASE_CHUNK_MULTIPLIER = 1;
-    static final String DEFAULT_SORTING_METHOD = "chunk";
+    private final Path inputPath;
+    private final Path outputPath;
+    private final int nThreads;
+    private static final Logger LOGGER = Logger.getLogger(ParallelExternalLongSorter.class.getName());
     //endregion
 
-    public ParallelExternalLongSorter(Path inputPath, Path outputPath, int nThreads, String sortingMethod) throws Exception {
+    public ParallelExternalLongSorter(Path inputPath, Path outputPath, int nThreads) {
+        this.inputPath = inputPath;
+        this.outputPath = outputPath;
+        this.nThreads = nThreads;
         //region pre-condition verification (and delete old output)
-        debug("either lying to you or verifying constructor args (enable assertions, add '-ea' in your JVM opts)");
+        LOGGER.info("either lying to you or verifying constructor args (enable assertions, add '-ea' in your JVM opts)");
         assert nThreads >= 1 : "must have at least 1 thread, not " + nThreads;
         assert !Files.isDirectory(inputPath) : "check yourself before you directoryour self";
         assert Files.isReadable(inputPath) : "input file is not readable";
-        var oldOutputDeleted = outputPath.toFile().delete();
-        debug((oldOutputDeleted ? "get that nasty mess out of here " : "it was like that when i got here!") + outputPath);
         validateOutputPath(outputPath);
         //endregion
+    }
 
-        debug("starting setup");
+    public void sort() throws IOException, InterruptedException, ExecutionException {
+        LOGGER.info("starting setup");
         Path tempFile = Files.createTempFile("external-sort-scratch-space", ".tmp");
         try (
                 FileChannel inputFileChannel = FileChannel.open(inputPath, Set.of(READ));
@@ -54,7 +61,7 @@ public class ParallelExternalLongSorter {
                 FileChannel outputFileChannel = FileChannel.open(outputPath, Set.of(CREATE, READ, WRITE))
         ) {
             //region make output files same size as input size
-            debug("JUST GIMME SOME ROOM TO BREATHE (preparing scratch space)");
+            LOGGER.info("JUST GIMME SOME ROOM TO BREATHE (preparing scratch space)");
             long inputSize = inputFileChannel.size();
             scratchFileChannel.truncate(inputSize);
             outputFileChannel.truncate(inputSize);
@@ -68,7 +75,7 @@ public class ParallelExternalLongSorter {
             }
 
             //region plan where to split the input file
-            debug("can longs get covid? better put them in pods just to be safe (preparing chunks)");
+            LOGGER.info("can longs get covid? better put them in pods just to be safe (preparing chunks)");
             int chunkCount = getChunkCount(nThreads, inputSize);
             Split[] splits = Split.createSplits(inputSize, chunkCount);
             //endregion
@@ -82,109 +89,40 @@ public class ParallelExternalLongSorter {
 
             //endregion
             //region create a pool to manage threads, give it tasks (chunk sorter), then wait for it to finish executing them all
-            debug("these sheets are so soft! just look at that thread count: " + nThreads);
-            debug("is this expired? its getting chunky (queueing chunk sort jobs)");
+            LOGGER.info("these sheets are so soft! just look at that thread count: " + nThreads);
+            LOGGER.info("is this expired? its getting chunky (queueing chunk sort jobs)");
             ExecutorService executor = Executors.newFixedThreadPool(nThreads);
             var futures = executor.invokeAll(List.of(chunkSorters));
-            debug("YOU'RE GONNA BE FRIENDS WHETHER YOU LIKE IT OR NOT! (joining threads)");
+            LOGGER.info("YOU'RE GONNA BE FRIENDS WHETHER YOU LIKE IT OR NOT! (joining threads)");
             executor.shutdown();  //tell the pool we're done giving it work
-            debug("GET TO THE CHOPPA! (awaiting threadpool termination)");
+            LOGGER.info("GET TO THE CHOPPA! (awaiting threadpool termination)");
             // blocks until all tasks complete
             boolean timedOut = !executor.awaitTermination(THREADPOOL_TIMEOUT_SECONDS, TimeUnit.SECONDS);
             if (timedOut) throw new RuntimeException("threadpool timed out");
-            for (var future: futures) future.get(); // propagate exceptions from child threads
+            for (var future : futures) future.get(); // propagate exceptions from child threads
             //endregion
 
             //region lock files, make scratch file chunks
             var scratchLock = scratchFileChannel.lock(0, scratchFileChannel.size(), false);
-            var outputLock= outputFileChannel.lock(0, outputFileChannel.size(), false);
-            var scratchChunks = new LongBuffer[(int)splits.length];
+            var outputLock = outputFileChannel.lock(0, outputFileChannel.size(), false);
+            var scratchChunks = new LongBuffer[(int) splits.length];
             for (int i = 0; i < splits.length; i++) {
                 scratchChunks[i] = scratchFileChannel.map(READ_ONLY, splits[i].bytePosition, splits[i].byteSize).asLongBuffer();
             }
             var outputBuffer = outputFileChannel.map(READ_WRITE, 0, inputSize).asLongBuffer();
             //endregion
             //region merge, unlock, make fun of java
-            debug("first element (pre-merge)" + outputBuffer.get(0));debug("last" + outputBuffer.get(outputBuffer.limit() - 1));
+            LOGGER.info("first element (pre-merge)" + outputBuffer.get(0));
+            LOGGER.info("last" + outputBuffer.get(outputBuffer.limit() - 1));
             merge(scratchChunks, outputBuffer);
-            debug("first element (post-merge)" + outputBuffer.get(0));debug("last" + outputBuffer.get(outputBuffer.limit() - 1));
+            LOGGER.info("first element (post-merge)" + outputBuffer.get(0));
+            LOGGER.info("last" + outputBuffer.get(outputBuffer.limit() - 1));
             outputLock.release();
             scratchLock.release();
-            debug("doing the world a favor and ending another java process (all done)");
+            LOGGER.info("doing the world a favor and ending another java process (all done)");
             //endregion
         }
     }
-
-    //region debug utils
-    private static void debug(String msg) {System.out.println("DEBUG [" + System.nanoTime() + "]: " + msg);}
-    private static void debug(Object msg) {System.out.println("DEBUG [" + System.nanoTime() + "]: " + msg);}
-    //endregion
-
-    public static void main(String[] args) throws Exception {
-        //region arg parsing
-        long start = System.currentTimeMillis();
-        debug(args);
-        String inputFileName = (args.length < 1) ? DEFAULT_INPUT_FILENAME : args[0];
-        String outputFileName = (args.length < 2) ? DEFAULT_OUTPUT_FILENAME : args[1];
-        final int nThreads = (args.length < 3) ? DEFAULT_NTHREADS : Integer.parseInt(args[2]);
-        String sortingMethod = (args.length < 4) ? DEFAULT_SORTING_METHOD : args[3];
-        final Path inputPath = Paths.get(inputFileName).toAbsolutePath();
-        final Path outputPath = Paths.get(outputFileName).toAbsolutePath();
-        debug("you put your long ints in " + inputPath);
-        debug("you take your long ints out" + outputPath);
-        debug("hardcore, " + DEFAULT_NTHREADS + " cores");
-        //endregion
-
-        //region test data generation
-        if (args.length >= 4) {
-            final int inputLength = Integer.parseInt(args[3]); // unit = # of long values
-            long generatorStart = System.nanoTime();
-            new DataFileGenerator(inputPath.toString(), inputLength).generate();
-            long generatorStop = System.nanoTime();
-            double generatorElapsedSeconds = (double) (generatorStop - generatorStart) / Math.pow(10, 9);
-            System.out.printf("Finished writing test file with %d long integers in %,.3f seconds%n",
-                    inputLength, generatorElapsedSeconds);
-        }
-        //endregion
-
-        new ParallelExternalLongSorter(inputPath, outputPath, nThreads, DEFAULT_SORTING_METHOD);
-        //region post-condition verification
-        try (
-                FileChannel inFC = FileChannel.open(inputPath, Set.of(READ));
-                FileChannel outFC = FileChannel.open(outputPath, Set.of(READ, WRITE))
-        ) {
-            assert (inFC.size() == outFC.size()) : "expected in size (" + inFC.size() + ") == out size (" + outFC.size() + ")";
-            var in = inFC.map(READ_ONLY, 0, inFC.size()).asLongBuffer();
-            var out = outFC.map(READ_ONLY, 0, outFC.size()).asLongBuffer();
-            assert isEmpty(in) == isEmpty(out) : "expected output to be all 0's only if input is all 0's";
-        }
-        //endregion
-
-        long stop = System.currentTimeMillis();
-
-        System.out.printf("Total execution time: %d ms%n", stop - start);
-    }
-
-    //region verification utils
-    private static void validateOutputPath(Path outputPath) {
-        debug("bb pls" + outputPath);
-        // cuz if the output file already exists and its already sorted, we can't tell if the program worked or
-        // if it just happened to not crash
-        assert !outputPath.toFile().exists() : "expected outputPath not to exist";
-        Path parent = outputPath.getParent();
-        debug("who's ya directory" + parent);
-        assert Files.exists(parent) : "parent directory of output path doesn't exist";
-        assert Files.isWritable(parent) : "parent directory of output path isn't writable";
-    }
-
-    private static boolean isSorted(LongBuffer lb) {
-        return IntStream.range(0, lb.limit() - 1).allMatch(i -> lb.get(i) <= lb.get(i + 1));
-    }
-
-    private static boolean isEmpty(LongBuffer lb) {
-        return IntStream.range(0, lb.limit()).allMatch(i -> lb.get(i) == 0);
-    }
-    //endregion
 
     //region chunk count (how many splits/chunks should we make?)
     /*
@@ -223,9 +161,9 @@ public class ParallelExternalLongSorter {
         }
 
         assert count > 0 : "integer overflow occurred while trying to get chunk count";
-        debug("Count Chunkula: " + count);
-        var MB = Math.pow(10,6);
-        debug("peak chonk (nThreads * MB / # of chunks): " + expectedPeakMemoryUse/MB);
+        LOGGER.info("Count Chunkula: " + count);
+        var MB = Math.pow(10, 6);
+        LOGGER.info("peak chonk (nThreads * MB / # of chunks): " + expectedPeakMemoryUse / MB);
         return count;
     }
     //endregion
@@ -243,7 +181,7 @@ public class ParallelExternalLongSorter {
         //  Maximum heap size: "Smaller of 1/4th of the physical memory or 1 GB"
         //      src: https://docs.oracle.com/javase/8/docs/technotes/guides/vm/gc-ergonomics.html
         var workingMemoryLimit = Runtime.getRuntime().maxMemory() - overhead;
-        debug("Pepperidge Farm remembers (MB): " + (workingMemoryLimit / MB));
+        LOGGER.info("Pepperidge Farm remembers (MB): " + (workingMemoryLimit / MB));
         return workingMemoryLimit;
     }
     //endregion
@@ -251,7 +189,7 @@ public class ParallelExternalLongSorter {
     // TODO: too slow
     private void merge(LongBuffer[] presortedChunks, LongBuffer output) {
         //region precondition verification
-        debug("applying coconut oil (verifying merge preconditions)");
+        LOGGER.info("applying coconut oil (verifying merge preconditions)");
         // check preconditions
         assert output.position() == 0 : "expected output buffer to start at position 0";  // nothing has been written to output yet
         // output buffer is same size as cumulative size of presortedChunks
@@ -267,7 +205,7 @@ public class ParallelExternalLongSorter {
         //endregion
 
         //region implementation
-        PriorityQueue<LongBuffer> minHeap = new PriorityQueue<>(new ChunkHeadComparator());
+        PriorityQueue<LongBuffer> minHeap = new PriorityQueue<>(new Utils.ChunkHeadComparator());
         // initially populate the heap with any non-empty chunk
         for (var chunk : presortedChunks) {
             if (chunk.hasRemaining()) minHeap.add(chunk);
@@ -286,79 +224,7 @@ public class ParallelExternalLongSorter {
         assert output.position() == output.limit() : "expected output buffer's position to be at limit"; // output is full
         assert Stream.of(presortedChunks).noneMatch(Buffer::hasRemaining) : "expected all chunks to be drained"; // all elements in scratch space written to output
         assert isSorted(output) : "expected output to be sorted";
-        debug("wow, the chunks are gone! you really can use coconut oil for everything (merge finished)");
+        LOGGER.info("wow, the chunks are gone! you really can use coconut oil for everything (merge finished)");
         //endregion
     }
-
-    //region merge util
-    private static class ChunkHeadComparator implements Comparator<LongBuffer> {
-        @Override
-        public int compare(LongBuffer left, LongBuffer right) {
-            return Long.compare(left.get(left.position()), right.get(right.position()));
-        }
-    }
-    //endregion
-
-    //region chunk sort
-    @SuppressWarnings("ClassCanBeRecord") // suppressed because colab notebook's version of java doesn't support 'record'
-    private static final class ChunkSorter implements Callable<Void> {
-        private final FileChannel inputFileChannel;
-        private final FileChannel scratchFileChannel;
-        private final Split split;
-        private final String sortingMethod;
-
-        private ChunkSorter(FileChannel inputFileChannel, FileChannel scratchFileChannel, Split split, String sortingMethod) {
-            this.inputFileChannel = inputFileChannel;
-            this.scratchFileChannel = scratchFileChannel;
-            this.split = split;
-            this.sortingMethod = sortingMethod;
-        }
-
-        @Override
-        public Void call() throws Exception {
-            var input = inputFileChannel.map(READ_ONLY, split.bytePosition, split.byteSize).asLongBuffer();
-            var scratch = scratchFileChannel.map(READ_WRITE, split.bytePosition, split.byteSize).asLongBuffer();
-            assert input.position() == scratch.position() : "expected chunk in/out to have same position";
-            assert input.limit() == scratch.limit() : "expected chunk in/out to have same limit";
-            debug("preparing outdated VCR references (sorting chunk)");
-            scratch.mark();
-            long[] tmp = new long[input.remaining()];
-            input.get(tmp);
-            if (sortingMethod == "parallel" ) {
-                Arrays.parallelSort(tmp);
-            }
-            else {
-                Arrays.sort(tmp);
-            }
-            scratch.put(tmp);
-            debug("be kind, rewind (finished sorting chunk, rewinding chunk buffer)");
-            scratch.reset();
-            return null;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (obj == this) return true;
-            if (obj == null || obj.getClass() != this.getClass()) return false;
-            var that = (ChunkSorter) obj;
-            return Objects.equals(this.inputFileChannel, that.inputFileChannel) &&
-                    Objects.equals(this.scratchFileChannel, that.scratchFileChannel) &&
-                    Objects.equals(this.split, that.split);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(inputFileChannel, scratchFileChannel, split);
-        }
-
-        @Override
-        public String toString() {
-            return "ChunkSorter[" +
-                    "inputFileChannel=" + inputFileChannel + ", " +
-                    "scratchFileChannel=" + scratchFileChannel + ", " +
-                    "split=" + split + ']';
-        }
-
-    }
-    //endregion
 }
